Recording.java: Deserialization does not create methods.
Deserializing a Recording class no longer attempts to create Method
reflected references. Therefore a Recording may be deserialized
even if the properties it holds test information on are not on
the classpath at the time of deserialization. Instead the Method
references are created when the Recording is played back.

Special handling of serialization for TestEvent is no longer needed.